FE/Qt: bugref:3409: Runtime UI: Fixing few more parfait warnings in Machine-view code, related to implicit type conversion.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107446 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMachineView.cpp
@@ -445,7 +445,9 @@ QSize UIMachineView::maximumGuestSize()
 
 void UIMachineView::updateViewport()
 {
-    uimachine()->viewportChanged(screenId(), contentsX(), contentsY(), visibleWidth(), visibleHeight());
+    uimachine()->viewportChanged(screenId(),
+                                 (ulong)contentsX(), (ulong)contentsY(),
+                                 (ulong)visibleWidth(), (ulong)visibleHeight());
 }
 
 #ifdef VBOX_WITH_DRAG_AND_DROP
@@ -880,7 +882,7 @@ void UIMachineView::sltPerformGuestResize(const QSize &toSize)
                                               uimachine()->isScreenVisible(screenId()),
                                               false /* change origin? */,
                                               0 /* origin x */, 0 /* origin y */,
-                                              size.width(), size.height(),
+                                              (ulong)size.width(), (ulong)size.height(),
                                               0 /* bits per pixel */,
                                               true /* notify? */);
             }
@@ -896,7 +898,7 @@ void UIMachineView::sltPerformGuestResize(const QSize &toSize)
                                               true /* enabled? */,
                                               false /* change origin? */,
                                               0 /* origin x */, 0 /* origin y */,
-                                              size.width(), size.height(),
+                                              (ulong)size.width(), (ulong)size.height(),
                                               0 /* bits per pixel */,
                                               true /* notify? */);
             }
@@ -942,7 +944,7 @@ void UIMachineView::sltPerformGuestResize(const QSize &toSize)
                                           uimachine()->isScreenVisible(screenId()),
                                           false /* change origin? */,
                                           0 /* origin x */, 0 /* origin y */,
-                                          size.width(), size.height(),
+                                          (ulong)size.width(), (ulong)size.height(),
                                           0 /* bits per pixel */,
                                           true /* notify? */);
             m_lastSizeHint = size;
@@ -1649,7 +1651,9 @@ void UIMachineView::takePausePixmapLive()
     screenShot.fill(0);
 
     /* Acquire screen-shot image: */
-    uimachine()->acquireScreenShot(screenId(), screenShot.width(), screenShot.height(), KBitmapFormat_BGR0, screenShot.bits());
+    uimachine()->acquireScreenShot(screenId(),
+                                   (ulong)screenShot.width(), (ulong)screenShot.height(),
+                                   KBitmapFormat_BGR0, screenShot.bits());
 
     /* Take the device-pixel-ratio into account: */
     const double dDevicePixelRatioActual = frameBuffer()->devicePixelRatioActual();
@@ -2170,8 +2174,8 @@ QSize UIMachineView::scaledForward(QSize size) const
     const double dDevicePixelRatioFormal = frameBuffer()->devicePixelRatio();
     const double dDevicePixelRatioActual = frameBuffer()->devicePixelRatioActual();
     if (!frameBuffer()->useUnscaledHiDPIOutput())
-        size = QSize(size.width() * dDevicePixelRatioActual, size.height() * dDevicePixelRatioActual);
-    size = QSize(size.width() / dDevicePixelRatioFormal, size.height() / dDevicePixelRatioFormal);
+        size = QSize((int)(size.width() * dDevicePixelRatioActual), (int)(size.height() * dDevicePixelRatioActual));
+    size = QSize((int)(size.width() / dDevicePixelRatioFormal), (int)(size.height() / dDevicePixelRatioFormal));
 
     /* Return result: */
     return size;
@@ -2182,7 +2186,7 @@ QSize UIMachineView::scaledBackward(QSize size) const
     /* Take the device-pixel-ratio into account: */
     const double dDevicePixelRatioFormal = frameBuffer()->devicePixelRatio();
     const double dDevicePixelRatioActual = frameBuffer()->devicePixelRatioActual();
-    size = QSize(size.width() * dDevicePixelRatioFormal, size.height() * dDevicePixelRatioFormal);
+    size = QSize((int)(size.width() * dDevicePixelRatioFormal), (int)(size.height() * dDevicePixelRatioFormal));
     if (!frameBuffer()->useUnscaledHiDPIOutput())
         size = QSize(size.width() / dDevicePixelRatioActual, size.height() / dDevicePixelRatioActual);
 
@@ -2205,7 +2209,7 @@ void UIMachineView::updateMousePointerPixmapScaling(QPixmap &pixmap, int &iXHot,
     if (dScaleFactor > 1.0)
     {
         /* Scale the pixmap up: */
-        pixmap = pixmap.scaled(pixmap.width() * dScaleFactor, pixmap.height() * dScaleFactor,
+        pixmap = pixmap.scaled((int)(pixmap.width() * dScaleFactor), (int)(pixmap.height() * dScaleFactor),
                                Qt::IgnoreAspectRatio, Qt::SmoothTransformation);
         iXHot *= dScaleFactor;
         iYHot *= dScaleFactor;
